leg feature: case-insensitive string literals
In a grammar, a string literal immediately followed by a lowercase "i"
will ignore case when it matches. So "begin"i or 'begin'i will match
input strings "begin", "BEGIN", "bEgIn", etc...

This is useful for languages with case-insensitive keywords, such as
SQL and Pascal. The keywords otherwise have to be written as a series
of character classes, e.g. [Bb][Ee][Gg][Ii][Nn], which is awkward and
hard to read and generates less efficient code.

Limitation: Only works with ASCII letters.